Update fitsio write method for empty headers
Change the fitsio write method to write multi extension FITS files with more than one empty HDU.
The code now uses the ignore_empty=True attribute the fitsio.FITS constructor.
This is the stable method now that is implemented in fitsio and replaces the older ignore_empty constructor that was added to the write() method in the 'feature-ignore-empty' branch.